GH Actions: fix apt-get dependency hell
The command as-it-was, is currently resulting in the following error:
```
You might want to run 'apt --fix-broken install' to correct these.
The following packages have unmet dependencies:
 libpcre3-dev : Depends: libpcre3 (= 2:8.44-2+ubuntu20.04.1+deb.sury.org+1) but 2:8.39-12build1 is to be installed
 libpcrecpp0v5 : Depends: libpcre3 (>= 2:8.41) but 2:8.39-12build1 is to be installed
 php7.2-cgi : Depends: libpcre3 (>= 2:8.41) but 2:8.39-12build1 is to be installed
 php7.2-cli : Depends: libpcre3 (>= 2:8.41) but 2:8.39-12build1 is to be installed
 php7.2-fpm : Depends: libpcre3 (>= 2:8.41) but 2:8.39-12build1 is to be installed
 php7.2-phpdbg : Depends: libpcre3 (>= 2:8.41) but 2:8.39-12build1 is to be installed
E: Unmet dependencies. Try 'apt --fix-broken install' with no packages (or specify a solution).
```

So, following the advise in the error message, I'm adding `apt --fix-broken install` to the workflows.
I'm also splitting the two commands in the "Install locales" step into separate steps to allow for easier debugging of where things actually go wrong.